Extract anonymous classes
