+++ merlin-rxjava2/src/main/java/com/novoda/merlin/rxjava2/MerlinFlowable.java
@@ -5,16 +5,9 @@ import android.content.Context;
 import com.novoda.merlin.Merlin;
 import com.novoda.merlin.MerlinBuilder;
 import com.novoda.merlin.NetworkStatus;
-import com.novoda.merlin.registerable.bind.Bindable;
-import com.novoda.merlin.registerable.connection.Connectable;
-import com.novoda.merlin.registerable.disconnection.Disconnectable;
 
 import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
-import io.reactivex.FlowableEmitter;
-import io.reactivex.FlowableOnSubscribe;
-import io.reactivex.annotations.NonNull;
-import io.reactivex.functions.Cancellable;
 
 public class MerlinFlowable {
 
@@ -32,41 +25,8 @@ public class MerlinFlowable {
         return createFlowable(merlin);
     }
 
-    private static Flowable<NetworkStatus.State> createFlowable(final Merlin merlin) {
-        return Flowable.create(new FlowableOnSubscribe<NetworkStatus.State>() {
-            @Override
-            public void subscribe(@NonNull final FlowableEmitter<NetworkStatus.State> emitter) throws
-                                                                                               Exception {
-                merlin.registerConnectable(new Connectable() {
-                    @Override
-                    public void onConnect() {
-                        emitter.onNext(NetworkStatus.State.AVAILABLE);
-                    }
-                });
-                merlin.registerDisconnectable(new Disconnectable() {
-                    @Override
-                    public void onDisconnect() {
-                        emitter.onNext(NetworkStatus.State.UNAVAILABLE);
-                    }
-                });
-                merlin.registerBindable(new Bindable() {
-                    @Override
-                    public void onBind(NetworkStatus networkStatus) {
-                        NetworkStatus.State current = networkStatus.isAvailable() ? NetworkStatus.State.AVAILABLE : NetworkStatus.State.UNAVAILABLE;
-                        emitter.onNext(current);
-                    }
-                });
-                emitter.setCancellable(new Cancellable() {
-                    @Override
-                    public void cancel() throws
-                                         Exception {
-                        merlin.unbind();
-                    }
-                });
-
-                merlin.bind();
-            }
-        }, BackpressureStrategy.BUFFER)
+    private static Flowable<NetworkStatus.State> createFlowable(Merlin merlin) {
+        return Flowable.create(new MerlinFlowableOnSubscribe(merlin), BackpressureStrategy.BUFFER)
                        .distinctUntilChanged();
     }
 }
